extract conversion from middleware to own file
